Isolate and relocate test CLI functionality
Extracts cli outside of browser suite in prep for upcoming WP-CLI integration tests
